+++ packages/app/src/screens/Home.tsx
@@ -1,16 +1,14 @@
 import React from 'react';
-import { View, Text, StyleSheet } from 'react-native';
+import { View, StyleSheet } from 'react-native';
 import Modalize from 'react-native-modalize';
-import { Feather } from '@expo/vector-icons';
-import { UserContext } from '../contexts/UserContext';
-import PaymentModal from '../components/PaymentModal';
+import Scanner from '../components/Scanner';
+import SuccessModal from '../components/SuccessModal';
 import Tickets from '../components/Tickets';
-import ScanButton from '../components/ScanButton';
-import { NavigationScreenProps } from 'react-navigation';
+import { UserContext } from '../contexts/UserContext';
 
-const Home = ({ navigation }: NavigationScreenProps) => {
+const Home = () => {
 	const { user } = React.useContext(UserContext);
-	const { firstName, _id: userId } = user;
+	const { _id: userId } = user;
 	const modalRef = React.useRef<Modalize>(null);
 
 	const handleModalOpen = () => {
@@ -19,18 +17,9 @@ const Home = ({ navigation }: NavigationScreenProps) => {
 
 	return (
 		<View style={styles.container}>
-			<View style={styles.welcomeContainer}>
-				<Text style={styles.welcomeText}>Hello, {firstName}.</Text>
-				<Feather
-					size={24}
-					name='settings'
-					onPress={() => navigation.navigate('Settings')}
-					color='#505050'
-				/>
-			</View>
+			<Scanner userId={userId} onSuccess={handleModalOpen} />
 			<Tickets {...{ userId }} />
-			<PaymentModal {...{ modalRef, userId }} />
-			<ScanButton onPress={handleModalOpen} />
+			<SuccessModal {...{ modalRef, userId }} />
 		</View>
 	);
 };
@@ -38,22 +27,7 @@ const Home = ({ navigation }: NavigationScreenProps) => {
 const styles = StyleSheet.create({
 	container: {
 		flex: 1,
-		// alignItems: 'center',
-		padding: 15
-	},
-	welcomeContainer: {
-		flexDirection: 'row',
-		marginVertical: 10,
-		justifyContent: 'space-between',
-		alignItems: 'center'
-	},
-	welcomeText: {
-		fontSize: 22,
-		fontWeight: 'bold',
-		color: '#505050'
-	},
-	welcomeAction: {
-		padding: 5
+		backgroundColor: '#E8E8E8'
 	}
 });
